requirements.txt
django==4.2.1
psycopg2-binary==2.9.6
django-environ==0.10.0
djangorestframework==3.14.0
Pillow==9.5.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
dj-rest-auth==4.0.0
facebook-sdk==3.1.0
google-api-core==2.11.0
google-api-python-client==2.86.0
